Changed overlap behaviour.  Default is "union", "error" is gone.
I decided I did not like the default error behaviour.  I understand the rationale of not wanting to surprise users and forcing
them to make an explicit choice.  At the same time, I don't like the fragility of the whole tool breaking on such a simple error
(which may not even be an error, i.e. the user may be fully aware that the language sets don't match).  So there is now no
default error, and to avoid surprise users are informed of non equal language sets with a message in --verbose mode.

I chose "union" as a default rather than "intersection" because I feel that throwing away data by default should be avoided.