Check tutorial using scriptdoctest until `beast`
It looks like the tutorial is fine and executable line-by-line up to
before the first call to `beast`. For the example lines after that, I
need to install beast first.

The tests were run with the most recent version of `scriptdoctest`
available from github.com/Anaphory/scriptdoctest, which includes a
patched version of scripttest. Run it using

    $ python -m scriptdoctest tutorial.rst
    [...]