Start separating config magic and core functions
Support only core functionality (reading one and only one config file)
on the cli, and make the actual parsing of one or multiple config
files (as needed by tests) raise exceptions if files are not found.

Permit passing files or strings or lists of files or strings or dicts
to the configparser, so we can use FileType in the argparser.

Given that file existence is now checked by the ArgumentParser, fix
the expected error codes in the CLI tests.

Move interpolation tests to a different file and skip them, so we have
a reminder for implementing the config magic in a standalone tool.