Reflect configalchemy in documentation
The new bits of documentation now mention that configalchemy can be
used for it, but beastling cannot.

This is not true yet (beastling still has a "+" in its ArgumentParser)
and the documentation of both tools should be separated and
cross-linked, but at least the documentation reflects the desired and
available API better now.