Hand error handling to the dictionary iterator
There is no way, even assuming valid CLDF, to not read the second row of the
forms.csv file to find whether the codecolumn exists. In that case, we might as
well not do anything, but expect the KeyError from trying to find code_column in
the iteration to be thrown and passed to the user.