There were calibrations of empty clades. Fix it.
The “Austronesian” clade in the configuration tests was empty. Instead, we now
have one item more in the data which actually shares a non-family clade with
another item (dal, which is Cushitic like aas) so we can use that to test things
using Glottolog clades.

This language had to be added to all applicable data files, trees, etc. for
consistency.