When parsing a tree from file-or-string, fail earlier.
When trying to get a tree from a configuration line, it will be either
a filename or a Newick string. However, a filename with a typo is
nearly always a valid Newick tree containing only one node – and such
a tree is never a useful monophyly or starting tree, so if there are
no brackets in the Newick string, assume something has gone wrong and
die.