Wrap errors only when called from the command line.
Do not catch-and-sys.exit errors when called from `python -m`, so that the it is
easier to debug (using pdb.pm(), eg.) errors from calling Beastling from the
command line.

This is achieved by defining which errors should be caught and turned into
various exit codes using `sys.exit()` – the default is that all exceptions be
caught, but `__main__` re-defines that variable to a class that is never raised,
and therefore does not catch most exceptions.

I am sure there is a better way to run this, and I would be happy to see it
implemented, but I have no intuition what it might be.