Make beastling.tree an attribute of TreePrior instead
Let the TreePrior object handle the tag that reflects its associated state node.
The config.treeprior object is handed through anyway, and making the tag local
to the prior will make it easier to write the different tree priors as
subclasses, which may throw errors when asked to add things to their states that
are not compatible.